cheat.c, cpu/mips: fix prototype mismatches that break 3DS / newlib builds
Two functions defined with int32_t/uint32_t parameter or return types
have header prototypes that use plain int/unsigned, which conflict on
toolchains where int32_t = long int (newlib, 3DS CTR, some embedded
ARM):

- src/cheat.h:    int cheat_menu(struct osd_bitmap *, int);
  src/cheat.c:    int32_t cheat_menu(struct osd_bitmap *, int32_t)
- src/cpu/mips/mips.h: extern unsigned mips_dasm(char *, unsigned);
  src/cpu/mips/mips.c:  unsigned mips_dasm(char *, uint32_t)

On Linux x86-64 LP64, int32_t = int, so the conflict was latent.  On
3DS CTR the build fails with:

    error: conflicting types for cheat_menu;
    have long int(struct osd_bitmap *, long int)

Reverts the definitions to match their headers.  Same compiled code
on the systems where they previously built clean.

This is the only place in the tree where my UINT*/INT* -> stdint sweep
changed a function signature visible across a header boundary.  All
other functions either are static, or had matching int32_t signatures
in both .h and .c, or are connected via tilemap_create-style function
pointers that already use uint32_t throughout.

Reported-by build CI on libretro-build-amd64-ubuntu (xenial gcc 9)
and miyoo arm32 toolchain.

+++ src/cpu/mips/mips.c
@@ -1219,11 +1219,11 @@ const char *mips_info( void *context, int regnum )
 	return "";
 }
 
-unsigned mips_dasm( char *buffer, uint32_t pc )
+unsigned mips_dasm( char *buffer, unsigned pc )
 {
 	unsigned ret;
 	change_pc32( pc );
-	sprintf( buffer, "$%08x", cpu_readop32( pc ) );
+	sprintf( buffer, "$%08x", (unsigned int)cpu_readop32( pc ) );
 	ret = 4;
 	change_pc32( NEXTPC );
 	return ret;
